- Fixed potential memory leaks in RegionCollection - Updated Locus and Region - updated Makefile to compile the new classes
git-svn-id: https://svn.ritchielab.psu.edu/svn/projects/biobin@1422 9e62ecf6-770d-0410-a9d7-e62a33e3d092

+++ trunk/src/knowledge/Region.cpp
@@ -65,7 +65,7 @@ void Region::addAliases(const string& aliases, const string& sep){
 		_aliases.insert(aliases.substr(init_pos, end_pos-init_pos));
 		init_pos = end_pos + sep.size();
 		end_pos = aliases.find(sep, init_pos);
-	} while(end_pos != string::npos);
+	} while((int) end_pos != (int) string::npos);
 }
